[ADM-887][frontend]: feat: useDefaultValues to make projectName accept imports.
