+++ frontend/src/containers/ConfigStep/index.tsx
@@ -1,9 +1,3 @@
-import { closeAllNotifications } from '@src/context/notification/NotificationSlice';
-import BasicInfo from '@src/containers/ConfigStep/BasicInfo';
-import { useAppDispatch } from '@src/hooks/useAppDispatch';
-import { ConfigStepWrapper } from './style';
-import { useLayoutEffect } from 'react';
-
 import {
   basicInfoSchema,
   boardConfigSchema,
@@ -14,19 +8,19 @@ import {
   IPipelineToolData,
   ISourceControlData,
 } from '@src/containers/ConfigStep/Form/schema';
-import {
-  basicInfoDefaultValues,
-  boardConfigDefaultValues,
-  pipelineToolDefaultValues,
-  sourceControlDefaultValues,
-} from '@src/containers/ConfigStep/Form/defaultValues';
+import { useDefaultValues } from '@src/containers/ConfigStep/Form/useDefaultValues';
+import { closeAllNotifications } from '@src/context/notification/NotificationSlice';
 import { SourceControl } from '@src/containers/ConfigStep/SourceControl';
 import { PipelineTool } from '@src/containers/ConfigStep/PipelineTool';
 import { selectConfig } from '@src/context/config/configSlice';
+import BasicInfo from '@src/containers/ConfigStep/BasicInfo';
+import { useAppDispatch } from '@src/hooks/useAppDispatch';
 import { useAppSelector } from '@src/hooks/useAppDispatch';
 import { Board } from '@src/containers/ConfigStep/Board';
 import { useForm, FormProvider } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
+import { ConfigStepWrapper } from './style';
+import { useLayoutEffect } from 'react';
 
 const ConfigStep = () => {
   const dispatch = useAppDispatch();
@@ -39,27 +33,28 @@ const ConfigStep = () => {
   const { isShow: isShowBoard } = configData.board;
   const { isShow: isShowPipeline } = configData.pipelineTool;
   const { isShow: isShowSourceControl } = configData.sourceControl;
+  const defaultValues = useDefaultValues();
 
   const basicInfoMethods = useForm<IBasicInfoData>({
-    defaultValues: basicInfoDefaultValues,
+    defaultValues: defaultValues.basicInfo,
     resolver: yupResolver(basicInfoSchema),
     mode: 'onChange',
   });
 
   const boardConfigMethods = useForm<IBoardConfigData>({
-    defaultValues: boardConfigDefaultValues,
+    defaultValues: defaultValues.boardConfig,
     resolver: yupResolver(boardConfigSchema),
     mode: 'onChange',
   });
 
   const pipelineToolMethods = useForm<IPipelineToolData>({
-    defaultValues: pipelineToolDefaultValues,
+    defaultValues: defaultValues.pipelineTool,
     resolver: yupResolver(pipelineToolSchema),
     mode: 'onChange',
   });
 
   const sourceControlMethods = useForm<ISourceControlData>({
-    defaultValues: sourceControlDefaultValues,
+    defaultValues: defaultValues.sourceControl,
     resolver: yupResolver(sourceControlSchema),
     mode: 'onChange',
   });
